requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
reportlab
python-multipart